+++ ChangeLog.md
@@ -1,3 +1,7 @@
+#2.2.10
+  * Fix detailed export of module reinforcement packages
+  * Use damagedist for exact breakdown of weapons that have more than one type of damage
+
 #2.2.9
   * Use SSL-enabled server for shortlinks
   * Add falloff for weapons